pyproject.toml
[project]
name = "mlops-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Bogachevv"}
]

readme = "README.md"
requires-python = ">=3.12,<3.15"

dependencies = [
    "torch (==2.9)",
    "torchvision (==0.24)",
    "transformers (<=4.55)",
    "peft (<=0.18)",
    "datasets (<=4.4)",
    "trl (==0.17)",
    "pandas (<=2.3)",
    "polars (<2.0)",
    "omegaconf",
    "hydra-core (==1.3)",
    "numexpr (>=2.14.1,<3.0.0)",
    "evaluate (>=0.4.6,<0.5.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "rouge-score (>=0.1.2,<0.2.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]

[dependency-groups]
test = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "codecov-cli",
    "pytest-cov (>=7.0.0, <8.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = true
packages = [{ include = "dialogue_summarization", from = "src" }, { include = "metrics", from = "src" }]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
python_files = ["test_*.py"]

[tool.coverage.report]
omit = ["**/__init__.py"]